+++ src/org/cicirello/search/operators/permutations/ScrambleMutation.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021  Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -38,11 +38,9 @@ import org.cicirello.math.rand.RandomIndexer;
  * the two end points.  
  * On average, a scramble mutation moves approximately n/3 elements.</p>
  *
- * @since 1.0
- *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.8.2020
+ * @version 4.13.2021
  */
 public class ScrambleMutation implements MutationOperator<Permutation> {